Ensure the typescript compiler doesn't recursively search for types.
The default behaviour of typescript is to recursively search for types in the `node_modules/@types` directory, then its parent `../node_modules/@types` and its parent `../../node_modules/@types` and so on and so forth.

This is probably fine in most cases but if you project is within a sub-directory of another project it can lead to very strange behaviours. I'm therefore disabling that search as thrift generated projects will be self contained anyway.